Fixed workers running out of resources, updated dependencies.
This was a whole ordeal—the inventory item JSON file from bungie must have increased enough in size to hit the cloudflare worker memory limit when requesting the data and then serializing it. As a dumb quick fix I created a new worker that is just responsible for handling inventory items. It seems to work, but is a mess 🤷‍♀️